+++ api/src/routes/partitionRoutes.ts
@@ -11,3 +11,4 @@ router.delete('/:id', PartitionController.delete);
 
 export default router;
 
+